+++ lib/appsignal.rb
@@ -19,6 +19,7 @@ module Appsignal
       require 'appsignal/integrations/delayed_job'
       require 'appsignal/integrations/sidekiq'
       require 'appsignal/integrations/resque'
+      require 'appsignal/integrations/sequel'
     end
 
     def load_instrumentations
@@ -151,10 +152,10 @@ module Appsignal
     #
     # @since 0.8.7
     def without_instrumentation
-      agent.paused = true if agent
+      Appsignal::Transaction.current.pause! if Appsignal::Transaction.current
       yield
     ensure
-      agent.paused = false if agent
+      Appsignal::Transaction.current.resume! if Appsignal::Transaction.current
     end
   end
 end
@@ -164,6 +165,10 @@ require 'appsignal/config'
 require 'appsignal/event_formatter'
 require 'appsignal/marker'
 require 'appsignal/native'
+require 'appsignal/rack/listener'
+require 'appsignal/rack/instrumentation'
+require 'appsignal/rack/sinatra_instrumentation'
+require 'appsignal/rack/js_exception_catcher'
 require 'appsignal/params_sanitizer'
 require 'appsignal/subscriber'
 require 'appsignal/transaction'
@@ -171,3 +176,4 @@ require 'appsignal/version'
 require 'appsignal/rack/listener'
 require 'appsignal/rack/instrumentation'
 require 'appsignal/integrations/rails'
+require 'appsignal/js_exception_transaction'