+++ lib/appsignal/integrations/rails.rb
@@ -5,22 +5,32 @@ if defined?(::Rails)
     module Integrations
       class Railtie < ::Rails::Railtie
         initializer 'appsignal.configure_rails_initialization' do |app|
-          app.middleware.insert_before(
-            ActionDispatch::RemoteIp,
-            Appsignal::Rack::Listener
-          )
+          Appsignal::Integrations::Railtie.initialize_appsignal(app)
         end
 
-        config.after_initialize do
+        def self.initialize_appsignal(app)
           # Start logger
           Appsignal.start_logger(Rails.root.join('log'))
 
           # Load config
           Appsignal.config = Appsignal::Config.new(
             Rails.root,
-            Rails.env,
+            ENV.fetch('APPSIGNAL_APP_ENV', Rails.env),
             :name => Rails.application.class.parent_name
           )
+
+          app.middleware.insert_before(
+            ActionDispatch::RemoteIp,
+            Appsignal::Rack::Listener
+          )
+
+          if Appsignal.config.active? &&
+            Appsignal.config[:enable_frontend_error_catching] == true
+            app.middleware.insert_before(
+              Appsignal::Rack::Listener,
+              Appsignal::Rack::JSExceptionCatcher,
+            )
+          end
 
           Appsignal.start
         end